chore: restructure types for reuse across workspaces
