+++ src/models/howto.models.tsx
@@ -1,6 +1,6 @@
 import { ISelectedTags } from './tags.model'
 import { DBDoc, IModerable } from './common.models'
-import type { IConvertedFileMeta } from '../components/ImageInput/ImageInput'
+import type { IConvertedFileMeta } from '../types'
 import type { IUploadedFileMeta } from '../stores/storage'
 
 /**
