+++ src/pages/User/content/SpaceProfile.tsx
@@ -34,7 +34,7 @@ import PVCIcon from 'src/assets/images/plastic-types/pvc.svg'
 // import V4MemberIcon from 'src/assets/icons/icon-v4-member.svg'
 
 import type { IUploadedFileMeta } from 'src/stores/storage'
-import type { IConvertedFileMeta } from 'src/components/ImageInput/ImageInput'
+import type { IConvertedFileMeta } from 'src/types'
 
 import { UserStats } from './UserStats'
 import UserContactAndLinks from './UserContactAndLinks'
