+++ src/stores/User/user.store.ts
@@ -10,7 +10,7 @@ import { auth, EmailAuthProvider, IFirebaseUser } from 'src/utils/firebase'
 import { RootStore } from '..'
 import { ModuleStore } from '../common/module.store'
 import { Storage } from '../storage'
-import { IConvertedFileMeta } from 'src/components/ImageInput/ImageInput'
+import type { IConvertedFileMeta } from 'src/types'
 import { formatLowerNoSpecial, randomID } from 'src/utils/helpers'
 import { logger } from 'src/logger'
 import { getLocationData } from 'src/utils/getLocationData'
