+++ src/stores/common/module.store.ts
@@ -5,7 +5,7 @@ import { ISelectedTags } from 'src/models/tags.model'
 import { IDBEndpoint, ILocation } from 'src/models/common.models'
 import { includesAll } from 'src/utils/filters'
 import { RootStore } from '../index'
-import { IConvertedFileMeta } from 'src/components/ImageInput/ImageInput'
+import type { IConvertedFileMeta } from 'src/types'
 import { IUploadedFileMeta, Storage } from '../storage'
 import { useCommonStores } from 'src/index'
 import { logger } from 'src/logger'
