+++ src/types/index.ts
@@ -1,0 +1,6 @@
+export interface IConvertedFileMeta {
+    photoData: Blob
+    objectUrl: string
+    name: string
+    type: string
+}
